Refactor: package 정리 #1
