+++ src/main/java/Calculator/handler/IO/InputHandler.java
@@ -1,4 +1,4 @@
-package Calculator;
+package Calculator.handler.IO;
 
 public interface InputHandler {
