+++ src/main/java/Calculator/handler/IO/SystemInputHandler.java
@@ -1,4 +1,4 @@
-package Calculator;
+package Calculator.handler.IO;
 
 import java.util.Scanner;
