+++ src/main/java/Calculator/handler/IO/SystemOutputHandler.java
@@ -1,4 +1,4 @@
-package Calculator;
+package Calculator.handler.IO;
 
 public class SystemOutputHandler implements OutputHandler {
