+++ src/main/java/Calculator/handler/calculate/ArithmeticCalculateHandler.java
@@ -1,10 +1,12 @@
-package Calculator;
+package Calculator.handler.calculate;
+
+import Calculator.operator.Operator;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
 
-import static Calculator.CalculateUtil.isDigit;
-import static Calculator.CalculateUtil.isOperator;
+import static Calculator.handler.calculate.CalculateUtil.isDigit;
+import static Calculator.handler.calculate.CalculateUtil.isOperator;
 
 public class ArithmeticCalculateHandler implements CalculateHandler {
